Changes: Added controler for polynomial update, restructed entities to hold floats instead of ints if needed

+++ src/main/java/pl/demo/polynomial/domain/Polynomial.java
@@ -21,8 +21,8 @@ public class Polynomial {
     @OneToMany(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
     private List<Variable> variables;
 
-    private int rangeStart;
+    private float rangeStart;
 
-    private int rangeEnd;
+    private float rangeEnd;
 
 }
